Function to create a hexagon with a given radius, center, and rotation

Function to create a hexagon with a given radius, center, and rotation

Usage

st_create_hexagon(center, radius, rotation = 0, crs = 2193)

Arguments

center

coordinates of the center of the hexagon

radius

radius of the hexagon

rotation

rotate corners

crs

crs for sf object

Value

A hexagon sf object

Examples

hexagon_sf <- create_hexagon(center = c(1600000, 5500000), radius = 100000, rotation = pi/8, crs = 2193)
#> Error in create_hexagon(center = c(1600000, 5500000), radius = 1e+05,     rotation = pi/8, crs = 2193): could not find function "create_hexagon"